Refactored Bucket views
