+++ app/bucket/views.py
@@ -1,5 +1,7 @@
-from flask import Blueprint, request, make_response, jsonify, abort
+from flask import Blueprint, request, abort
 from app.auth.helper import token_required
+from app.bucket.helper import response, response_for_created_bucket, response_for_user_bucket, get_response, \
+    get_user_bucket_json_list
 from app.models import User, Bucket
 from app import db
 from sqlalchemy import exc
@@ -70,7 +72,7 @@ def get_bucket(current_user, bucket_id):
         else:
             if user_bucket:
                 return response_for_user_bucket(user_bucket.json())
-            return response_for_user_bucket({})
+            return response_for_user_bucket([])
     except ValueError:
         return response('failed', 'Please provide a valid Bucket Id', 400)
 
@@ -100,6 +102,12 @@ def edit_bucket(current_user, bucket_id):
 @bucket.route('/bucketlists/<bucket_id>', methods=['DELETE'])
 @token_required
 def delete_bucket(current_user, bucket_id):
+    """
+    Deleting a User Bucket from the database if it exists.
+    :param current_user:
+    :param bucket_id:
+    :return:
+    """
     try:
         int(bucket_id)
         try:
@@ -133,69 +141,3 @@ def handle_400_errors(e):
     :return:
     """
     return response('failed', 'Bad Request', 400)
-
-
-def response_for_user_bucket(user_bucket):
-    """
-    Return the response for when a single bucket when requested by the user.
-    :param user_bucket:
-    :return:
-    """
-    return make_response(jsonify({
-        'status': 'success',
-        'bucket': user_bucket
-    }))
-
-
-def response_for_created_bucket(user_bucket, status_code):
-    """
-    Method returning the response when a bucket has been successfully created.
-    :param status_code:
-    :param user_bucket: Bucket
-    :return: Http Response
-    """
-    return make_response(jsonify({
-        'status': 'success',
-        'id': user_bucket.id,
-        'name': user_bucket.name,
-        'createdAt': user_bucket.create_at,
-        'modifiedAt': user_bucket.modified_at
-    })), status_code
-
-
-def response(status, message, code):
-    """
-    Helper method to make a http response
-    :param status: Status message
-    :param message: Response message
-    :param code: Response status code
-    :return: Http Response
-    """
-    return make_response(jsonify({
-        'status': status,
-        'message': message
-    })), code
-
-
-def get_user_bucket_json_list(user_buckets):
-    """
-    Make json objects of the user buckets and add them to a list.
-    :param user_buckets: Bucket
-    :return:
-    """
-    buckets = []
-    for user_bucket in user_buckets:
-        buckets.append(user_bucket.json())
-    return buckets
-
-
-def get_response(buckets):
-    """
-    Make a http response for BucketList get requests.
-    :param buckets: Bucket
-    :return:
-    """
-    return make_response(jsonify({
-        'status': 'success',
-        'buckets': buckets
-    })), 200
